Corregida estructura de ficheros a un proyecto maven
