+++ rchavarria/src/main/java/Basket.groovy
@@ -1,7 +1,25 @@
 
 class Basket {
+	private static DISCOUNT_FACTORS = [0, 0, 0.05, 0.1, 0.2, 0.25]
 
+	private BookListUtil bookList = new BookListUtil()
+	
 	def price(def books) {
-		0
+		if(!books) return 0
+		
+		def discount = computeDiscount(books)
+		
+		(1 - discount) * 8 * books.size()
+	}
+	
+	private computeDiscount(def books){
+		def discount = 0
+		
+		if(bookList.areAllDifferent(books)){
+			def discountFactor = books.size()
+			discount = DISCOUNT_FACTORS[discountFactor]
+		}
+		
+		return discount
 	}
 }
