+++ rchavarria/src/main/java/BookListUtil.groovy
