+++ rchavarria/src/main/java/BookListUtilTest.groovy
